refactor: use typescript
add packages
