+++ package.json
@@ -1,22 +1,33 @@
 {
-    "name": "ws_task",
+    "name": "ws_task-remote-control",
     "version": "1.0.0",
-    "description": "Template for the WebSocket remote control task",
+    "description": "Made from Template for the WebSocket remote control task",
     "main": "index.js",
     "type": "module",
     "scripts": {
-        "start": "node ./index.js",
-        "start:dev": "nodemon ./index.js",
-        "test": "echo \"Error: no test specified\" && exit 1"
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "start": "ts-node src/index.ts",
+        "start:dev": "npx eslint ./src && ts-node src/index.ts",
+        "dev": "nodemon run start:dev"
     },
-    "author": "Andrei Auchynnikau",
+    "author": "Andrei Auchynnikau / Aless133",
     "keywords": [],
     "license": "ISC",
     "devDependencies": {
-        "nodemon": "^2.0.16"
+        "@types/node": "^18.11.18",
+        "@types/ws": "^8.5.3",
+        "@typescript-eslint/eslint-plugin": "^5.47.1",
+        "@typescript-eslint/parser": "^5.47.1",
+        "eslint": "^8.30.0",
+        "eslint-config-prettier": "^8.5.0",
+        "eslint-plugin-prettier": "^4.2.1",
+        "nodemon": "^2.0.20",
+        "prettier": "^2.8.1",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.4"        
     },
     "dependencies": {
         "@nut-tree/nut-js": "^2.3.0",
-        "ws": "^8.8.0"
+        "ws": "^8.11.0"
     }
 }
